+++ lib/src/visitor/async_evaluate.dart
@@ -575,14 +575,21 @@ final class _EvaluateVisitor
     ];
 
     var metaModule = BuiltInModule("meta",
-        functions: [...meta.global, ...meta.local, ...metaFunctions],
+        functions: [...meta.moduleFunctions, ...metaFunctions],
         mixins: metaMixins);
 
     for (var module in [...coreModules, metaModule]) {
       _builtInModules[module.url] = module;
     }
 
-    functions = [...?functions, ...globalFunctions, ...metaFunctions];
+    functions = [
+      ...?functions,
+      ...globalFunctions,
+      ...[
+        for (var function in metaFunctions)
+          function.withDeprecationWarning('meta')
+      ]
+    ];
     for (var function in functions) {
       _builtInFunctions[function.name.replaceAll("_", "-")] = function;
     }