+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.78.0
+
+* `@import` is now officially deprecated, as are global built-in functions that
+  are available within built-in modules. See [the Sass blog post] for more
+  details on the deprecation process.
+
+[the Sass blog]: https://sass-lang.com/blog/import-is-deprecated/
+
 ## 1.77.8
 
 * No user-visible changes.