+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 10.5.0
+
+* No user-visible changes.
+
 ## 10.4.8
 
 * No user-visible changes.